Move the profile info from App.js to Navigation.js
