+++ packages/react-frontend/src/pages/ProfilePage.js
@@ -1,4 +1,6 @@
-import { Typography, Stack } from "@mui/material";
+import { useState, useEffect } from "react";
+import { Box, Typography, Stack } from "@mui/material";
+import { jwtDecode } from "jwt-decode";
 
 const ProfilePage = (props) => {
   return (
